PyGFX: Offset image/volume transforms.
It seems logical to me that if you have an image of size (100,100) in a
scene, the bounding box of that rectangle (untransformed) should be ((0,
0), (100, 100)). PyGFX transforms images/volumes such that the center of
image[0, 0] is at (0, 0). This results in the bounding box being ((-0.5,
-0.5), (99.5, 99.5)).

This change corrects the bounding boxes by offsetting any transform
applied to pygfx images/volumes.