Vispy Offset image/volume transforms (#23)
* PyGFX: Offset image/volume transforms.

It seems logical to me that if you have an image of size (100,100) in a
scene, the bounding box of that rectangle (untransformed) should be ((0,
0), (100, 100)). PyGFX transforms images/volumes such that the center of
image[0, 0] is at (0, 0). This results in the bounding box being ((-0.5,
-0.5), (99.5, 99.5)).

This change corrects the bounding boxes by offsetting any transform
applied to pygfx images/volumes.

* Offset vispy transforms instead.

@tlambert03 and @marktsuchida noted that centering the the samples at
the integer locations is more correct.

My goal was mostly just to ensure uniformity between our two adaptors -
I wasn't thinking about correctness :)

* Transform tests

* Use determine_backend for test skips

This gracefully handles the case where both are installed

* Fix tests

---------

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>