Repurpose _snx_get_native for native widget
All NodeAdaptors had a _snx_get_native method that was used to return
the native pygfx/vispy object, however there wasn't really a
(production) use case for this. There is, however, a production use case
for using that same method on the CanvasAdaptors to get the native GUI
widget. This commit massages the existing (non-production) uses of
_snx_get_native to use other means, and repurposes that name for just
the canvases