Remove View.blending
With PyGfx v0.13.0, they removed blending mode from the renderer in
favor of alpha parameters on each material in the scene. This aligns
more closely with how VisPy does things, which is important for us
because it ignored View.blending anyways!

All of this is to say that I think it's worth (a) removing View.blending
for now to avoid the errors and (b) coming up with a new mechanism for
capturing transparency within the model (when we have a use for it?)