look_at: Set the near plane much closer
Previously, the near clipping plane was set to the distance from the
closest point to the scene from the camera. This isn't good enough,
because that distance can become smaller when the camera orbits, causing
annoying clipping.

A near plane of 1 is much better, although we may want to decrease it
further later.