Enabling maintaining aspect ratio as view changes
I'm not super happy with this change so I'm going to leave it on the ndv
branch for now, but we need a way to set the aspect ratio of the camera,
and then a way to maintain that aspect when the view layout changes.
This is one way to do that. There are other additional behaviors that
aren't yet possible (such as zooming in/out to maintain view boundaries)
as the view layout changes, but I haven't yet figured out a good way to
do that.